added master scraper run script with options + readout coloring

+++ DeepDives_Run.py
@@ -2,42 +2,54 @@
 import subprocess
 import time
 import os
-import json
+import sys
 from drgmissions_scraper_utils import (
+    yes_or_no,
     upload_file,
     enable_system_time,
     wait_until_next_thursday_11am_utc,
     get_previous_thursday_date,
     kill_process_by_name_starts_with,
     maximize_window,
+    minimize_window,
     subprocess_wrapper,
+    wrap_with_color,
     print,
     cfg
 )
 
 def main():
+    kill_process_by_name_starts_with('FSD')
+    kill_process_by_name_starts_with('Unreal')
+    auto = True
+    if len(sys.argv) > 1:
+        if sys.argv[1].strip() == 'manual':
+            auto = False
+
     for f in os.listdir():
         if f.startswith('DD_') and f.endswith('.json'):
             os.remove(f)
             
     maximize_window()
-    print(os.getcwd())
     
     time_service_query = subprocess_wrapper(['sc', 'query', 'w32time'], print_=False)()
     if 'RUNNING' not in time_service_query:
-        print('Enabling automatic system time...')
+        print(wrap_with_color('Enabling automatic system time...', '0;33'))
         enable_system_time()
         time.sleep(2)
     else:
-        subprocess_wrapper(['w32tm', '/resync'])()
-        
-    wait_until_next_thursday_11am_utc()
+        print(wrap_with_color('Resyncing system time in case of drift...','0;33'))
+        subprocess_wrapper(['w32tm', '/resync'], print_=False)()
+    
+    if auto:
+        wait_until_next_thursday_11am_utc()
 
     with open('./mods/mods.txt', 'w') as f:
         f.write('DeepDivesScraper : 1')
         f.close()
         
     subprocess.Popen(['start', 'steam://run/548430//'], shell=True)
+    minimize_window('FSD-Win64-Shipping.exe')
 
     files = []
     start_time = time.time()
@@ -66,19 +78,22 @@ def main():
                 
         if files:
             for file in files:
-                upload_file(cfg, file)
-                os.remove(file)
+                if auto:
+                    upload_file(cfg, file)
+                    os.remove(file)
+                elif yes_or_no('Upload JSON? Y/N: '):
+                    upload_file(cfg, file)
             break
         
         time.sleep(0.5)
 
     with open('./mods/mods.txt', 'w') as f:
         f.close()
-
-    subprocess.Popen(['shutdown', '/s', '/f', '/t', '0'])
+        
+    if auto:
+        subprocess.Popen(['shutdown', '/s', '/f', '/t', '0'])
 
 try:
     main()
 except Exception as e:
     print(e)
-    input('Press enter to exit...')
